Split Backend into multiple files following a sort of MVC pattern. There's now a controller and service layer
