+++ backend/src/config/serverConfig.js
@@ -1,0 +1,5 @@
+
+// Holds Environment Args for the Server instance
+module.exports = ({
+    PORT_NUMBER: 8080
+})
