+++ backend/src/config/spotifyConfig.js
@@ -1,0 +1,6 @@
+
+// Holds information required to communicate with spotify API
+module.exports = ({
+    SPOTIFY_ID: '7d4e98d5926f47fd887b33c0db094c72',
+    SPOTIFY_CLIENT_SECRET: '5274969d76bb4006964b92076acc6194'
+})
