+++ backend/src/controller/controller.js
@@ -1,0 +1,7 @@
+/*
+  API FLOW: 
+    - user searches an artist name, server return list of artists
+    - user confirms the artist selection, server returns a list of albums
+    - user selects a list of albums, server returns each track from those albums.
+*/
+
